Support polytonic Greek diacritics in the Greek keyboard
Although modern Greek is typically written with only acute accent marks,
polytonic diacritics are still used by the Greek orthodox church, the
Greek newspaper Estia, and students of classical Greek texts. Supporting
these special characters is helpful to all of those groups.

Note that for the short vowels (epsilon, upsilon, iota, and omicron) all
of the diacritics permitted on lowercase letters are also permitted on
their uppercase forms. However, for the long vowels (alpha, eta, and
omega) the diacritics permitted on uppercase letters are a subset of the
diacritics permitted on lowercase letters. That's just how Greek is.

Test: Manually turned on the Greek (Greece) keyboard in the aosp_x86-eng
emulator and tried out all the new characters with Shift, with Shift
Lock, and without any modifiers.

Bug: 142007431
Change-Id: Iec4a82cab02aea5176721e8447f7dfdba46bacb3
Signed-off-by: ghou8s <ghou8s@gmail.com>